requirements.txt
asgiref==3.6.0
Django==4.2
django-credit-cards==0.4.1
Pillow==9.5.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
